fix: try tmp

+++ src/index.js
@@ -10,23 +10,29 @@ const sleepSync = (ms) => {
 }
 
 const main = () => {
+
+    // Get the inputs
+    const echoResults = core.getInput("echo-results");
+    const startTcpdump = core.getInput("start-tcpdump");
+
     // Check if dns.pcap exists
-    if (!fs.existsSync("dns.pcap")) {
+    if (startTcpdump && !fs.existsSync("dns.pcap")) {
 
         const command = 'sudo';
-        const args = ['tcpdump', '-n', '-i', 'any', '-w', 'dns.pcap', 'port', '53'];
+        const args = ['tcpdump', '-n', '-i', 'any', '-w', 'tmp/dns.pcap', 'port', '53'];
 
         const tcpdumpProcess = spawn(command, args, {
             detached: true, // Detach the child process from the parent
             stdio: 'ignore', // Ignore stdin, stdout, and stderr
         });
 
+        // Let tcpdump run for 2 seconds to get started
         sleepSync(2000);
 
         // Unref the child process to allow the parent process to exit
         tcpdumpProcess.unref();
 
-    } else {
+    } else if (echoResults && fs.existsSync("tmp/dns.pcap")) {
         exec('sudo pkill tcpdump', (err, stdout, stderr) => {
             if (err) {
                 console.log(err);
@@ -40,9 +46,11 @@ const main = () => {
         sleepSync(5000);
 
         // Convert PCAP to JSON
-        const packets = pcap_parser.parsePcapFile("dns.pcap");
+        const packets = pcap_parser.parsePcapFile("tmp/dns.pcap");
         console.log(packets);
-    };
+    } else {
+        console.log("No DNS packets capture started.");
+    }
 }
 
 main();
